+++ CHANGELOG.md
@@ -9,7 +9,10 @@ with respect to its command line interface and HTTP interface.
 ## [Unreleased](//github.com/opentable/sous/compare/0.5.85...HEAD)
 ### Added
 * Server: more flexible, agile logging API.
+* Server: logging API includes adding context fields to child loggers.
+* All: updated to enable better running of tests in teamcity
 ### Changed
+* Client: `sous deploy` and `sous newdeploy` now synonyms. Expect `sous newdeploy` to be removed.
 * Client: Display more information in case timeout of sous newdeploy.  Also show Executor Message if failed deploy.
 * Server: cleanups to logging output